fix : fix type compatible error

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: rjeong <rjeong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/09 17:57:54 by rjeong            #+#    #+#             */
-/*   Updated: 2022/07/09 18:13:19 by rjeong           ###   ########.fr       */
+/*   Updated: 2022/07/09 21:51:48 by rjeong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*(((unsigned char *) s) + i) && i < n)
